requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
Pillow==10.4.0
pytesseract==0.3.13
reportlab==4.2.5
